Moved thermal comps
