--- openconcept/components/__init__.py
@@ -1,1 +1,0 @@
-from .manifold import FlowSplit, FlowCombine
