+++ openconcept/thermal/tests/test_manifold.py
@@ -3,7 +3,7 @@ import unittest
 import numpy as np
 from openmdao.utils.assert_utils import assert_near_equal, assert_check_partials
 from openmdao.api import IndepVarComp, Group, Problem
-from openconcept.components import FlowSplit, FlowCombine
+from openconcept.thermal import FlowSplit, FlowCombine
 
 
 class FlowSplitTestCase(unittest.TestCase):
