+++ openconcept/thermal/tests/test_thermal_comps.py
@@ -3,7 +3,7 @@ import unittest
 import numpy as np
 from openmdao.utils.assert_utils import assert_near_equal, assert_check_partials
 from openmdao.api import Problem, NewtonSolver, DirectSolver
-import openconcept.components.thermal as thermal
+import openconcept.thermal as thermal
 
 class PerfectHeatTransferCompTestCase(unittest.TestCase):
     """
