+++ openconcept/utilities/math/tests/test_add_subtract_comp.py
@@ -5,7 +5,6 @@ import unittest
 import numpy as np
 
 from openmdao.api import Problem, Group, IndepVarComp
-#from openmdao.components.add_subtract_comp import AddSubtractComp
 from openconcept.utilities.math.add_subtract_comp import AddSubtractComp
 from openmdao.utils.assert_utils import assert_near_equal, assert_check_partials
